Move install binaries function
