+++ apps/vs-code-designer/src/app/commands/binaries/resetValidateAndInstallBinaries.ts
@@ -1,0 +1,9 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See LICENSE.md in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+import type { IActionContext } from '@microsoft/vscode-azext-utils';
+
+export const resetValidateAndInstallBinaries = async (activateContext: IActionContext) => {
+  console.log(activateContext);
+};
