+++ apps/vs-code-designer/src/onboarding.ts
@@ -2,8 +2,9 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
+import { validateAndInstallBinaries } from './app/commands/binaries/validateAndInstallBinaries';
 import { activateAzurite } from './app/utils/azurite/activateAzurite';
-import { promptInstallBinariesOption, validateAndInstallBinaries } from './app/utils/binaries';
+import { promptInstallBinariesOption } from './app/utils/binaries';
 import { promptStartDesignTimeOption } from './app/utils/codeless/startDesignTimeApi';
 import { runWithDurationTelemetry } from './app/utils/telemetry';
 import { getGlobalSetting } from './app/utils/vsCodeConfig/settings';
